Move tic tac toe model to game in models folder and update import statement
