+++ spec/tic_tac_toe.spec.js
@@ -1,4 +1,4 @@
-import TicTacToe from 'tic_tac_toe';
+import TicTacToe from 'app/models/game';
 import Board from 'app/models/board';
 import Player from 'app/models/player';
